server/requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.22
mysqlclient==2.0.1
protobuf==3.14.0
PyMySQL==0.10.1
six==1.15.0
Werkzeug==1.0.1
